requirements.txt
streamlit
google-genai
groq
mistralai
python-dotenv
python-pptx
python-docx
pypdf
